requirements.txt
numpy
Pillow
python-gdcm
pydicom
typer[all]
